==> src/data-wrangling/01_rfmo_cleaning_iccat.Rmd <==
---
title: "RFMO Data Cleaning - ICCAT"
output: html_notebook
---

This is a notebook designed to grab data saved in Google Drive and re-format it for easier analyses later.

ICCAT data were mostly in Microsoft Access databases, so I had to export the queries into csv files for import.

# Table of Contents: {#top}

- [T2CE Web]
- [Combining Files]

```{r, include=FALSE}
# Import libraries
library(tidyverse)
library(readxl)
library(reshape)
library(here)
```

# Housekeeping
```{r, echo=TRUE}
# List files to choose from
file_path <- file.path(here::here(), "data-updated/rfmo-data/inputs/iccat/")
save_location <- file.path(here::here(), "data-updated/rfmo-data/intermediates/iccat/")
outputs_location <- file.path(here::here(), "data-updated/rfmo-data/outputs/")
```

# Load Catch Data
```{r, echo=TRUE}
# MDB files first... 
# mdb.get is acting weird with the current link, saving to downloads locally
# mdb_names <- Hmisc::mdb.get(file = "/Users/echelleburns/Downloads/t2ce_20220131web.mdb", tables = TRUE)
# 
# for(i in mdb_names) {
#   temp_get <- Hmisc::mdb.get(file = "/Users/echelleburns/Downloads/t2ce_20220131web.mdb",
#                              tables = i)
#   assign(i, temp_get)
# }
# 
# mdb_data <- t2ce %>%
#   # left_join(., Catalogue) %>%
#   # left_join(., DatasetTypes) %>%
#   left_join(., Flags %>% dplyr::select(FleetID, FlagName)) %>%
#   left_join(., TimeStrata %>%
#               dplyr::select(TimeStrataID, Description) %>%
#               dplyr::rename(TimePeriodID = TimeStrataID,
#                      time_period = Description))
# 
# # Save it
# write.csv(mdb_data, file.path(file_path, "t2ce_20220131web.csv"), row.names = F)

# Load t2ce web
t2ce_web <-  read.csv(file.path(file_path, "t2ce_20220131web.csv")) %>% 
  filter((DSetTypeID %in% c(".w", "n.") & CatchUnit %in% c("kg", "nr")) | (DSetTypeID == "nw" & CatchUnit == "nr"))
```

# Load Coded Files
```{r, echo=TRUE}
# Load gear codes
# Currently used gear
gear <- read_xls(paste0(file_path, "CODES_Gears.xls"), skip=2)
gear <- gear[,c(1:3)]

# Historically used gear
historic_gear <- read_xls(paste0(file_path, "CODES_Gears.xls"), skip=51, col_names = colnames(gear))

# Combine gears
gear <- rbind(gear, historic_gear)
remove(historic_gear)

# Gear group codes
gear_groups <- read_xls(paste0(file_path, "CODES_Gears.xls"), skip=2)
gear_groups <- gear_groups[, c("Code", "Name")] %>% distinct_all()

# Load species codes
species <- read_xlsx(paste0(file_path, "CODES_Species.xlsx"), skip=1)

# Load time period codes
time_periods <- read_xls(paste0(file_path, "CODES_TimePeriods.xls"), skip=2)

# Load effort types
effort_types <- read_xls(paste0(file_path, "CODES_EffortTypes.xls"), skip=2)
```

# T2CE Web
From the readme: This document provides a summarized description of the Task II Catch and Effort statistics (T2CE) published on the ICCAT website and also explains the changes in structure of the information published. In addition, it briefly describes the applications available to extract T2CE information. The information published (pertaining to 1950 through to 2007) is only the T2CE reported by the CPCs to ICCAT (Contracting and Cooperating non-Contracting Parties, Entities or Fishing Entities) and some other non-Cooperating Parties.
```{r, echo=TRUE}
# Keep only data from the year 2012 onward
t2ce_web <- t2ce_web %>% 
  filter(YearC >= 2012)

# Pivot tables so each row represents a different species
t2ce_web <- t2ce_web %>% 
  pivot_longer(cols=BFT:oSks, names_to = 'SpeciesCode', values_to = 'MonthlyCatch') %>% 
  filter(MonthlyCatch > 0)

# Reform data so different effort types are in different rows
t2ce_web_eff1 <- t2ce_web[,-c(16:17)]
t2ce_web_eff2 <- t2ce_web[,-c(14:15)]
colnames(t2ce_web_eff1)[14:15] <- c("Effort", "EffortType")
colnames(t2ce_web_eff2)[14:15] <- c("Effort", "EffortType")
t2ce_web <- rbind(t2ce_web_eff1, t2ce_web_eff2)
remove(t2ce_web_eff1, t2ce_web_eff2)

# Join with other datasets
t2ce_web <- t2ce_web %>% 
  select(-time_period) %>% 
  left_join(gear_groups, by=c("GearGrpCode" = "Code")) %>%
  left_join(effort_types[,c("EffortTypeCode", "EffortTypeName")], by=c("EffortType" = "EffortTypeCode")) %>%
  left_join(species[,c("SpeciesCode", "ScieName", "English")], by="SpeciesCode") %>% 
  left_join(time_periods, by="TimePeriodID")

# Fix some effort values
t2ce_web <- t2ce_web %>% 
  mutate(EffortTypeName = ifelse(is.na(EffortTypeName), EffortType, EffortTypeName),
         EffortType = toupper(EffortType),
         EffortTypeName = ifelse(EffortType == "NO.NETS", "number of nets", EffortTypeName),
         EffortTypeName = ifelse(EffortType == "NO.MTZAS", "number of MTZAS", EffortTypeName),
         EffortTypeName = ifelse(EffortType == "NO.FADS.VIS", "number of FAD vis", EffortTypeName), 
         EffortTypeName = ifelse(EffortType == "KM.SETS", "kilometers of set", EffortTypeName),
         EffortTypeName = ifelse(EffortType == "HOURS.SEA", "hours at sea", EffortTypeName))

# Add names for species that are not in the species list
t2ce_web <- t2ce_web %>% 
  mutate(English = ifelse(SpeciesCode == "oSks", "sharks nei", 
                          ifelse(SpeciesCode == "oTun", "tunas nei", English)))

# Make all names uppercase
t2ce_web$ScieName<- toupper(t2ce_web$ScieName)
t2ce_web$English<- toupper(t2ce_web$English)

# Fix latlon coordinates based on QuadID
# From readme file -- 1: NE, 2:SE, 3: SW, 4:NW
t2ce_web <- t2ce_web %>% 
  filter(Lat <= 90) %>%
  mutate(Lon = ifelse(QuadID == 3 | QuadID == 4, Lon*(-1), Lon), 
         Lat = ifelse(QuadID == 2 | QuadID == 3, Lat*(-1), Lat))

# Format relevant columns
t2ce_web <- t2ce_web %>% 
  mutate(rfmo = "ICCAT", 
         year = YearC, 
         month = TimePeriodID, 
         time_period = TPeriodGroup,
         country = FlagName, 
         gear = Name, 
         settype = SchoolTypeCode, 
         longitude = Lon,
         latitude = Lat, 
         effort = Effort, 
         effort_units = EffortTypeName, 
         species_commonname = English, 
         species_sciname = ScieName, 
         catch = MonthlyCatch, 
         catch_units = CatchUnit, 
         spatial_notes = paste0("center of ", SquareTypeCode, " cell"), 
         quality_code = NA, 
         was_generated = "no", 
         sources = "t2ce_20220131web")

# Keep only relevant columns
t2ce_web <- t2ce_web[,c("rfmo", "year", "month", "time_period", "country", "gear", "settype", "longitude", "latitude", "effort", "effort_units", "species_commonname", "species_sciname", "catch", "catch_units", "spatial_notes", "quality_code", "was_generated", "sources")]

# Make column values consistent with other datasets 
# Countries
t2ce_web  <- t2ce_web %>% 
  mutate(country = gsub("C\xf4te d'Ivoire", "Côte d'Ivoire", country), 
         country = gsub("U.S.A.", "United States", country),
         country = gsub("Espa\xf1a", "España", country), 
         country = gsub("China PR", "China", country), 
         country = gsub("Korea Rep.", "Korea", country), 
         country = gsub("Guin\xe9e Rep.", "Guinée Rep.", country),
         country = gsub("Cura\xe7ao", "Curaçao", country), 
         country = gsub("UK|EU|FR|[.]", "", country), 
         country = gsub("&", "and", country), 
         country = gsub("-", "", country),
         )

# Gears
t2ce_web  <- t2ce_web %>% 
  mutate(gear = gsub("purse seine", "purse-seine", gear), 
         gear = gsub("rood & reel", "rod-and-reel", gear), 
         gear = gsub("U", "u", gear), 
         gear = gsub("trammerl", "trammel", gear)
         )

# Settype
t2ce_web  <- t2ce_web %>% 
  mutate(settype = gsub("n/a", NA, settype), 
         settype = gsub("FSC", "free schools", settype))

# Effort units
t2ce_web  <- t2ce_web %>% 
  mutate(effort_units = gsub("Number of ", "", effort_units), 
         effort_units = gsub("hooks used", "hooks", effort_units), 
         effort_units = gsub("days fishing", "fishing days", effort_units),
         effort_units = gsub("Line days", "fishing days ", effort_units),
         effort_units = gsub(" \\(.*\\)", "", effort_units), 
         effort_units = gsub("used", "", effort_units), 
         effort_units = gsub("-", " ", effort_units), 
         effort_units = gsub("P", "p", effort_units), 
         effort_units = gsub("made", "", effort_units), 
         effort_units = gsub("sets", "number of sets", effort_units),
         effort_units = str_trim(as.character(effort_units), side="both"), 
         effort_units = gsub("none", NA, effort_units), 
         effort_units = ifelse(effort_units == "", NA, effort_units)
         )

# Months and time periods
t2ce_web <- t2ce_web %>% 
  mutate(month = gsub(17, NA, month), 
         month = gsub(13, 1, month), 
         month = gsub(14, 4, month), 
         month = gsub(15, 7, month), 
         month = gsub(16, 10, month),
         month = as.numeric(month),
         time_period = tolower(time_period),
         time_period = gsub("month", "monthly", time_period), 
         time_period = gsub("trimester", "quarterly", time_period))

# Catch units (and catches to convert kg to metric tonnes)
# 1 metric tonne = 1000 kg 
t2ce_web <- t2ce_web %>% 
  mutate(catch_units = gsub("nr", "count", catch_units), 
         catch = ifelse(catch_units == "kg", catch/1000, catch), 
         catch_units = gsub("kg", "metric tonnes", catch_units)
         )

# Generate data
# Generate quarterly data (aggregates)
quart_agg <- t2ce_web %>% 
  filter(time_period == "monthly") %>%
  mutate(month = ceiling(month/3),
         month = ifelse(month == 1, 1, # fix so each quarter == first month of quarter
                        ifelse(month == 2, 4, 
                               ifelse(month == 3, 7, 10))),
         time_period = "quarterly", 
         was_generated = "yes") %>% 
  group_by_at(setdiff(names(t2ce_web), c("catch", "effort"))) %>%
  summarise(catch = sum(catch), 
            effort = sum(effort))

# Generate yearly data (aggregates)
yearly_agg <- t2ce_web %>% 
  filter(time_period == "monthly") %>%
  mutate(month = NA,
         time_period = "yearly", 
         was_generated = "yes") %>% 
  group_by_at(setdiff(names(t2ce_web), c("catch", "effort"))) %>%
  summarise(catch = sum(catch), 
            effort = sum(effort))

# Combine raw with generated data
t2ce_web <- t2ce_web %>% 
  bind_rows(quart_agg) %>% 
  bind_rows(yearly_agg)

# Save File
filename <- "t2ce_web.csv"
write.csv(t2ce_web, paste0(save_location, filename), row.names = FALSE)

# Clean up
remove(t2ce_web, quart_agg, yearly_agg)
```
[Back To Top](#top)

# Combining Files
```{r, echo=TRUE}
# Find list of cleaned data
files_list <- list.files(save_location, pattern="csv")

# Make sure it doesn't include the 'all' dataset
files_list <- files_list[!grepl("all", files_list)]

# Dataframe to hold combined data
combined_dat <- NULL

# Loop through files and save to master dataframe
for(file in files_list) {
  temp <- read.csv(paste0(save_location, file))
  combined_dat <- rbind(combined_dat, temp)
}

# Remove instances of NA for lat or lon (some issues with grid in sfskh)
combined_dat <- combined_dat %>% 
  filter(!is.na(latitude) & !is.na(longitude))

# Filter so only 1x1 or 5x5 data are included
combined_dat <- combined_dat %>% 
  filter(spatial_notes %in% c("center of 5x5 cell", "center of 1x1 cell") & 
           grepl("purse-seine|purse seine|longline|long line|long-line", gear))

# Save final data file for this RFMO
filename <- "iccat-all.csv"
write.csv(combined_dat, paste0(outputs_location, filename), row.names = FALSE)
```
[Back To Top](#top)
