==> src/data-wrangling/01_rfmo_cleaning_wcpfc.Rmd <==
---
title: "RFMO Data Cleaning - WCPFC"
output: html_notebook
---

This is a notebook designed to grab data saved in Google Drive and re-format it for easier analyses later.

# Table of Contents: {#top}

- [Bycatch Longline]
- [Bycatch Purse Seine]
- [Longline by Flag and Year]
- [Purse Seine by Flag and Year]
- [Combining Files]

```{r, include=FALSE}
# Import libraries
library(tidyverse)
library(reshape)
library(readxl)
library(here)
```

# Housekeeping
```{r, echo=TRUE}
# List files to choose from
file_path <- file.path(here::here(), "data-updated/rfmo-data/inputs/wcpfc/")
save_location <- file.path(here::here(), "data-updated/rfmo-data/intermediates/wcpfc/")
outputs_location <- file.path(here::here(), "data-updated/rfmo-data/outputs/")
```

# Load Data
```{r, echo=TRUE}
# Read in files
country_codes_ref <- read.csv(file.path(here::here(), "data-updated", "two_letter_country_codes.csv"))
  
# Bycatch files
bycatch_longline <- read_xlsx(paste0(file_path, "BDEP Tables (MASTER - 27 July 2021).xlsx"), sheet = "Table 4")
bycatch_seine <- read_xlsx(paste0(file_path, "BDEP Tables (MASTER - 27 July 2021).xlsx"), sheet = "Table 5")

# Target fish files 
list_files <- list.files(file_path, pattern ="CSV")
for(file in list_files) { 
    name <- as.character(str_split(file, "[.]")[[1]][1])
    temp <- read.csv(paste0(file_path, file))
    assign(name, temp)
}

remove(temp)

# Read in species file (bycatch data only)
species <- read_xlsx(paste0(file_path, "BDEP Tables (MASTER - 27 July 2021).xlsx"), sheet = "Species List")

# ASFIS species list
asfis_species <- read.delim(paste0(file_path, "ASFIS_sp_2019.txt"), sep=",")

# Read in yearly observed effort file
effort <- read_xlsx(paste0(file_path, "BDEP Tables (MASTER - 27 July 2021).xlsx"), sheet = "Table 1", skip=1)
effort <- effort[,c(1:5)]
colnames(effort) <- c("year", "fishery", "total_effort", "total_observed_effort", "observer_coverage")
effort <- effort[!is.na(effort$total_effort), ]
```

# Bycatch Longline
```{r, echo=TRUE}
# Rename columns for ease of later merges (lots of spaces and parantheses)
colnames(bycatch_longline) <- c("year", "gear", "latitude", "longitude", "species_cat", "species_grp", "num_vessels_obs", "catch", "obs_catchrate", "obs_mortality", "obs_mortalityrate", "obs_release" )

# Join with other datasets
bycatch_longline <- left_join(bycatch_longline %>% 
                                mutate(species_grp = ifelse(species_grp == "OTHERS", 
                                                            paste0(species_cat, "-", species_grp), species_grp)), 
                              species[,c("BDEP/WCPFC Species Grouping", "common_name", "scientific_name")],
                              by=c("species_grp" = "BDEP/WCPFC Species Grouping")) %>% 
  mutate(common_name = ifelse(is.na(common_name) & species_grp == "SHK-OTHERS", "SHARKS NEI", common_name), 
         effort = catch/obs_catchrate*1000)

bycatch_longline <- left_join(bycatch_longline, effort %>% 
                                filter(fishery == "L") %>% 
                                select(year, observer_coverage), by=c("year"))

# Make all species names all caps
bycatch_longline$species_sciname <- toupper(bycatch_longline$scientific_name)
bycatch_longline$species_commonname <- toupper(bycatch_longline$common_name)

# Add columns that are not currently here
bycatch_longline <- bycatch_longline %>% 
  mutate(rfmo = "WCPFC", 
         month = NA,
         time_period = "yearly",
         country = NA,
         gear = "longline",
         settype = NA, 
         effort_units = "hooks",
         catch_units = "count",
         spatial_notes = "center of 5x5 cell", 
         was_generated = "no", 
         quality_code = NA, 
         sources = "bycatch and effort from observers")

# Select relevant columns 
bycatch_longline <- bycatch_longline[,c("rfmo", "year", "month", "time_period", "country", "gear", "settype", "longitude", "latitude", "effort", "effort_units", "species_commonname", "species_sciname", "catch", "catch_units", "spatial_notes", "quality_code", "was_generated", "sources")]

# Save File
filename <- "bycatch_longline.csv"
write.csv(bycatch_longline, paste0(save_location, filename), row.names = FALSE)

# Clean up
remove(bycatch_longline)
```
[Back To Top](#top)

# Bycatch Purse Seine
```{r, echo=TRUE}
# Rename columns for ease of later merges (lots of spaces and parantheses)
colnames(bycatch_seine) <- c("year", "gear", "latitude", "longitude", "species_cat", "species_grp", "num_vessels_obs", "effort", "catch", "obs_catchrate", "obs_mortality", "obs_mortalityrate", "obs_release" )

# Join with other datasets
bycatch_seine <- left_join(bycatch_seine  %>% 
                             mutate(species_grp = ifelse(species_grp == "OTHERS", 
                                                            paste0(species_cat, "-", species_grp), species_grp)),
                           species[,c("BDEP/WCPFC Species Grouping", "common_name", "scientific_name")],
                           by=c("species_grp" = "BDEP/WCPFC Species Grouping")) %>% 
  mutate(common_name = ifelse(is.na(common_name) & species_grp == "SHK-OTHERS", "SHARKS NEI", common_name))

# Join with observer coverage
bycatch_seine <- left_join(bycatch_seine, effort %>% 
                                filter(fishery == "S") %>% 
                                select(year, observer_coverage), by=c("year"))

# Make all species names all caps
bycatch_seine$species_sciname <- toupper(bycatch_seine$scientific_name)
bycatch_seine$species_commonname <- toupper(bycatch_seine$common_name)

# Add columns that are not currently here
bycatch_seine <- bycatch_seine %>% 
  mutate(rfmo = "WCPFC", 
         month = NA,
         time_period = "yearly",
         country = NA,
         gear = "purse-seine",
         settype = NA,  
         effort_units = "number of sets",
         catch_units = "count",
         spatial_notes = "center of 5x5 cell", 
         quality_code = NA, 
         was_generated = "no", 
         sources = "bycatch and effort from observers")

# Select relevant columns 
bycatch_seine <- bycatch_seine[,c("rfmo", "year", "month", "time_period", "country", "gear", "settype", "longitude", "latitude", "effort", "effort_units", "species_commonname", "species_sciname", "catch", "catch_units", "spatial_notes", "quality_code", "was_generated", "sources")]

# Save File
filename <- "bycatch_seine.csv"
write.csv(bycatch_seine, paste0(save_location, filename), row.names = FALSE)

# Clean up
remove(bycatch_seine)
remove(effort)
```
[Back To Top](#top)

# Longline by Flag and Year
```{r, echo=TRUE}
# Use an easier name
ll_flagyr <- WCPFC_L_PUBLIC_BY_FLAG_YR
remove(WCPFC_L_PUBLIC_BY_FLAG_YR)

# Only keep data from year 2012+
ll_flagyr <- ll_flagyr %>% 
  filter(yy >= 2012)

# Update flags
country_codes <- ll_flagyr %>% 
  select(flag_id) %>% 
  distinct_all() %>% 
  filter(flag_id != "") %>% 
  left_join(country_codes_ref %>% dplyr::rename(country = Name), 
            by = c("flag_id" = "Code"))

# Merge flag codes
ll_flagyr <- left_join(ll_flagyr, country_codes, by="flag_id")

# Fix the lat/long columns
# First convert N/S, E/W to pos and negative
# The readme: Latitude == latitude of south-west corner; Longitude == longitude of south-west corner for 5x5 degree grids
# So, to get the center, we need to add 2.5 degrees in each direction
ll_flagyr <- ll_flagyr %>% 
  separate(lat_short, into = c('latitude', 'lat_direction'), sep = 2) %>% 
  separate(lon_short, into = c('longitude', 'lon_direction'), sep = 3) %>% 
  mutate(latitude = ifelse(lat_direction == "S", as.numeric(latitude)*(-1)+2.5, as.numeric(latitude)+2.5), 
         longitude = ifelse(lon_direction == "W", as.numeric(longitude)*(-1)+2.5, as.numeric(longitude)+2.5))

# Pivot tables so each row represents a different species
ll_flagyr <- ll_flagyr %>% 
  pivot_longer(cols=alb_c:oth_n, names_to = 'species_code', values_to = 'catch') %>% 
  filter(catch > 0)

# Break species_code column into species name and catch units
ll_flagyr <- ll_flagyr %>% 
  separate(species_code, into=c("species_code", "catch_units"), sep=3) %>% 
  mutate(catch_units = gsub("_n", "count", catch_units), 
         catch_units = gsub("_c", "metric tonnes", catch_units))

# Join with AFSIS species codes
ll_flagyr$species_code <- toupper(ll_flagyr$species_code)
ll_flagyr <- left_join(ll_flagyr, asfis_species[,c("X3A_CODE", "Scientific_name", "English_name")], by=c("species_code" = "X3A_CODE"))

# Make all species names all caps
ll_flagyr$species_sciname <- toupper(ll_flagyr$Scientific_name)
ll_flagyr$species_commonname <- toupper(ll_flagyr$English_name)

# Fix names where species code is "OTH" (other species)
ll_flagyr <- ll_flagyr %>% 
  mutate(species_sciname = ifelse(species_code == "OTH", "OTHER SPECIES", species_sciname), 
         species_commonname = ifelse(species_code == "OTH", "OTHER SPECIES", species_commonname))

# Add columns that are not currently here
ll_flagyr <- ll_flagyr %>% 
  mutate(rfmo = "WCPFC", 
         year = yy, 
         month = NA, 
         time_period = "yearly",
         gear = "longline",
         settype = NA,  
         effort = hhooks*100,
         effort_units = "hooks",
         spatial_notes = "center of 5x5 cell", 
         quality_code = NA, 
         was_generated = "no", 
         sources = "self-reported catch and effort by flag + year")

# Select relevant columns 
ll_flagyr <- ll_flagyr[,c("rfmo", "year", "month", "time_period", "country", "gear", "settype", "longitude", "latitude", "effort", "effort_units", "species_commonname", "species_sciname", "catch", "catch_units", "spatial_notes", "quality_code", "was_generated", "sources")]

# Save File
filename <- "longline_flagyear.csv"
write.csv(ll_flagyr, paste0(save_location, filename), row.names = FALSE)

# Clean up
remove(ll_flagyr)
```
[Back To Top](#top)

# Purse Seine by Flag and Year
```{r, echo=TRUE}
# Use an easier name
ps_flagyr <- WCPFC_S_PUBLIC_BY_FLAG_YEAR
remove(WCPFC_S_PUBLIC_BY_FLAG_YEAR)

# Only keep data from year 2012+
ps_flagyr <- ps_flagyr %>% 
  filter(yy >= 2012)

# Merge flag codes
ps_flagyr <- left_join(ps_flagyr, country_codes, by="flag_id")

# Fix the lat/long columns
# First convert N/S, E/W to pos and negative
# The readme: Latitude == latitude of south-west corner; Longitude == longitude of south-west corner for 5x5 degree grids
# So, to get the center, we need to add 2.5 degrees in each direction
ps_flagyr <- ps_flagyr %>% 
  separate(lat_short, into = c('latitude', 'lat_direction'), sep = 2) %>% 
  separate(lon_short, into = c('longitude', 'lon_direction'), sep = 3) %>% 
  mutate(latitude = ifelse(lat_direction == "S", as.numeric(latitude)*(-1)+2.5, as.numeric(latitude)+2.5), 
         longitude = ifelse(lon_direction == "W", as.numeric(longitude)*(-1)+2.5, as.numeric(longitude)+2.5))

# Pivot tables so each row represents the set type and the number of sets for that set type
ps_flagyr <- ps_flagyr %>% 
  pivot_longer(cols=sets_una:sets_oth, names_to = 'settype', values_to = 'effort') %>% 
  mutate(settype = gsub("sets_una", "unassociated school", settype), 
         settype = gsub("sets_log", "natural log/debris", settype), 
         settype = gsub("sets_dfad", "drifting FAD", settype), 
         settype = gsub("sets_afad", "anchored FAD", settype), 
         settype = gsub("sets_oth", "other", settype), 
         effort_units = "number of sets")

# Copy the data, but instead put number of days as the effort unit
ps_flagyr <- rbind(ps_flagyr, ps_flagyr %>% 
                      mutate(effort = days, 
                             effort_units = "fishing days"))

# Pivot tables so each row represents a different species
ps_flagyr <- ps_flagyr %>% 
  pivot_longer(cols=skj_c_una:oth_c_oth, names_to = 'species_code', values_to = 'catch') %>% 
  filter(catch > 0)

# Break species_code column into species name and catch units
ps_flagyr <- ps_flagyr %>% 
  separate(species_code, into=c("species_code", "catch_units", "catch_set"), sep=c(3,6)) %>% 
  mutate(catch_units = gsub("_c_", "metric tonnes", catch_units), 
         catch_set = gsub("una", "unassociated school", catch_set), 
         catch_set = gsub("log", "natural log/debris", catch_set), 
         catch_set = gsub("dfad", "drifting FAD", catch_set), 
         catch_set = gsub("afad", "anchored FAD", catch_set), 
         catch_set = gsub("oth", "other", catch_set))

# Only keep rows of catch that correspond to the settype == catch_set
ps_flagyr <- ps_flagyr %>% 
  filter(settype == catch_set)

# Join with ASFIS species codes
ps_flagyr$species_code <- toupper(ps_flagyr$species_code)
ps_flagyr <- left_join(ps_flagyr, asfis_species[,c("X3A_CODE", "Scientific_name", "English_name")], by=c("species_code" = "X3A_CODE"))

# Make all species names all caps
ps_flagyr$species_sciname <- toupper(ps_flagyr$Scientific_name)
ps_flagyr$species_commonname <- toupper(ps_flagyr$English_name)

# Fix names where species code is "OTH" (other species)
ps_flagyr <- ps_flagyr %>% 
  mutate(species_sciname = ifelse(species_code == "OTH", "OTHER SPECIES", species_sciname), 
         species_commonname = ifelse(species_code == "OTH", "OTHER SPECIES", species_commonname))

# Add columns that are not currently here
ps_flagyr <- ps_flagyr %>% 
  mutate(rfmo = "WCPFC", 
         year = yy, 
         month = NA, 
         time_period = "yearly",
         gear = "purse-seine",
         spatial_notes = "center of 5x5 cell", 
         quality_code = NA, 
         was_generated = "no", 
         sources = "self-reported catch and effort by flag + year")

# Select relevant columns 
ps_flagyr <- ps_flagyr[,c("rfmo", "year", "month", "time_period", "country", "gear", "settype", "longitude", "latitude", "effort", "effort_units", "species_commonname", "species_sciname", "catch", "catch_units", "spatial_notes", "quality_code", "was_generated", "sources")]

# Save File
filename <- "purseseine_flagyear.csv"
write.csv(ps_flagyr, paste0(save_location, filename), row.names = FALSE)

# Clean up
remove(ps_flagyr)
```
[Back To Top](#top)

# Combining Files
```{r, echo=TRUE}
# Find list of cleaned data
files_list <- list.files(save_location, pattern="csv")

# Make sure it doesn't include the 'all' dataset
files_list <- files_list[!grepl("all", files_list)]

# Dataframe to hold combined data
combined_dat <- NULL

# Loop through files and save to master dataframe
for(file in files_list) {
  temp <- read.csv(paste0(save_location, file))
  combined_dat <- combined_dat %>% bind_rows(temp)
}
# Remove instances of NA for lat or lon (some issues with grid in sfskh)
combined_dat <- combined_dat %>% 
  filter(!is.na(latitude) & !is.na(longitude))

# Filter so only 1x1 or 5x5 data are included
combined_dat <- combined_dat %>% 
  filter(spatial_notes %in% c("center of 5x5 cell", "center of 1x1 cell") & 
           grepl("purse-seine|purse seine|longline|long line|long-line", gear))

# Save final data file for this RFMO
filename <- "wcpfc-all.csv"
write.csv(combined_dat, paste0(outputs_location, filename), row.names = FALSE)
```
[Back To Top](#top)
