==> src/data-wrangling/03_species_list_for_sdms.Rmd <==
---
title: "Generate Species List"
output: html_notebook
---

This notebook grabs the unique species in the RFMO observer-reported data so that we can match SDMs to them later 

```{r}
# Load packages
library(tidyverse)
library(readxl)
library(rfishbase) # just for nei species
library(here)

###
# Load data
###

# rfmo data
all_data <- vroom::vroom(file.path(here::here(), "data-updated/rfmo-data/outputs/all_data.csv"))

# # Species groupings
# species_groups <- read_xlsx(file.path(here::here(),"data-updated/species-information/species_groups.xlsx", sheet = "species_list")) %>% 
#   filter(group %in% c("tunas", "tuna-like species", "sharks and rays")) %>% 
#   mutate(scientific_name = str_to_upper(scientific_name))
```

```{r}
# Only keep relevant data at useable resolutions
short_data <- all_data %>% 
  filter(spatial_notes %in% c("center of 5x5 cell", "center of 1x1 cell") &
             was_generated == "no") %>% 
  select(species_commonname, species_sciname, species_group) %>% 
  distinct_all()

# Some data cleaning
# short_data <- short_data %>% 
#   mutate(species_sciname = case_when(species_sciname == "TETRAPTURUS AUDAX" ~ "KAJIKIA AUDAX",
#                                      species_sciname == "MAKAIRA INDICA" ~ "ISTIOMPAX INDICA", 
#                                      species_sciname == "TETRAPTURUS ALBIDUS" ~ "KAJIKIA ALBIDA",
#                                      species_sciname == "TETRAPTURUS PFLUEGERI" ~ "TETRAPTURUS PFLUEGERI",
#                                      TRUE ~ species_sciname))


# Now standardize nei species if can...
short_data <- short_data %>%
  mutate(species_sciname = gsub(" SPP", "", species_sciname)) %>% 
  mutate(species_resolution = case_when(grepl("NEI|OTHER|SPP", species_commonname) ~ "NEI",
                                        grepl("NEI|OTHER|SPP", species_sciname) ~ "NEI",
                                        species_commonname == "SILKY OR BLACKTIP SHARK" ~ "NEI",
                                        species_commonname == "FRIGATE AND BULLET TUNAS" ~ "NEI",
                                        species_commonname == "FRIGATE TUNA AND BULLET TUNA" ~ "NEI",
                                        
                                        !grepl(" ", species_sciname) ~ "NEI",
                                        TRUE ~ "SPECIES SPECIFIC"))

# Some edits because we have some super families and suborders and rfishbase can't account for that
get_species_lists <- short_data %>%
  filter(species_resolution == "NEI" & species_sciname != "OTHER SPECIES") %>%
  select(species_sciname) %>%
  distinct_all() %>%
  mutate(species_sciname = str_to_sentence(species_sciname),
         species_sciname = ifelse(species_sciname == "Xiphioidei", "Xiphiidae", species_sciname), # Xiphioidei family
         species_sciname = ifelse(species_sciname == "Scombroidei", "Gempylidae", species_sciname), #Scombroidei family
         species_sciname = ifelse(species_sciname == "Thunnini", "Allothunnus", 
                                  species_sciname)) %>% #Thunnini family
  bind_rows(data.frame("species_sciname" = c("Istiophoridae", # Xiphioidei family
                                             "Scombridae", # Scombroidei family
                                             "Scombrolabracidae", #Scombroidei family
                                             "Trichiuridae", #Scombroidei family,
                                             "Carcharhiniformes", #sharks nei order
                                             "Heterodontiformes", #sharks nei order
                                             "Hexanchiformes", #sharks nei order
                                             "Lamniformes", #sharks nei order
                                             "Orectolobiformes", #sharks nei order
                                             "Pristiophoriformes", #sharks nei order
                                             "Squatiniformes", #sharks nei order
                                             "Auxis", #Thunnini family
                                             "Euthynnus", #Thunnini family
                                             "Katsuwonus", #Thunnini family
                                             "Thunnus" #Thunnini family
                                             )))

# function to iterate through options and return the results
get_names <- function(x) {
  subfam <- species_list(Subfamily = x)
  fam <- species_list(Family = x)
  gen <- species_list(Genus = x)
  cls <- species_list(Class = x)
  ord <- species_list(Order = x)

  final <- c(subfam, fam, gen, cls, ord)

  if(length(final) > 0) {
  return(final)
  }
}

# run for species list
species_list <- sapply(get_species_lists$species_sciname, get_names)

# turn into a dataframe for SDMs later
species_list <- data.frame(t(sapply(species_list,c))) %>%
  pivot_longer(colnames(.), names_to = "species_sciname") %>%
  mutate(value = gsub('"|c[(]|[)]', "", value)) %>%
  separate_rows(value, sep = ", ") %>%
  filter(value != "NULL") %>%
  rowwise() %>%
  mutate(species_sciname = as.character(species_sciname),
         species_sciname = case_when(species_sciname %in% c("Carcharhiniformes", #sharks nei order
                                             "Heterodontiformes", #sharks nei order
                                             "Hexanchiformes", #sharks nei order
                                             "Lamniformes", #sharks nei order
                                             "Orectolobiformes", #sharks nei order
                                             "Pristiophoriformes", #sharks nei order
                                             "Squaliformes", #sharks nei order
                                             "Squatiniformes" #sharks nei order
                                             ) ~ "Sharks nei",
                                     species_sciname %in% c("Istiophoridae", "Xiphiidae") ~ "Xiphioidei",
                                     species_sciname %in% c("Scombridae", "Gempylidae", "Scombrolabracidae", "Trichiuridae") ~ "Scombroidei",
                                     species_sciname %in% c("Allothunnus", "Auxis", "Euthynnus", "Katsuwonus", "Thunnus") ~ "Thunnini",
                                     TRUE ~ species_sciname),
         
           species_sciname = gsub("[.]1", "", species_sciname),
           species_sciname = str_to_upper(species_sciname)) %>%
  ungroup() %>%
  mutate(value = gsub("\n", "", value))

species_list <- species_list %>%
  bind_rows(data.frame("species_sciname" = c("CARCHARHINUS FALCIFORMIS/CARCHARHINUS LIMBATUS",
                                             "CARCHARHINUS FALCIFORMIS/CARCHARHINUS LIMBATUS", 
                                             "AUXIS THAZARD AND AUXIS ROCHEI", 
                                             "AUXIS THAZARD AND AUXIS ROCHEI"),
                       "value" = c("Carcharhinus falciformis", "Carcharhinus limbatus", 
                                   "Auxis thazard", "Auxis rochei"))) %>% 
  distinct_all()

species_threatened_fishing <- read.csv(file.path(here::here(), "data-updated/species-information/spp_list_fishing_threat.csv"))

species_list_for_sdms <- short_data %>%
  select(species_commonname, species_sciname, species_resolution, species_group) %>%
  distinct_all() %>%
  left_join(species_list) %>%
  mutate(value = ifelse(is.na(value), str_to_sentence(species_sciname), value)) %>%
  left_join(species_threatened_fishing %>%
              select(scientific_name) %>%
              mutate(threatened = "yes") %>%
              rename(value = scientific_name)) %>%
  # filter(species_group %in% c("tunas", "tuna-like species") | # remove for now
  filter(species_group == "sharks and rays" & threatened == "yes") %>%
  select(-threatened)

write.csv(species_list_for_sdms, file.path(here::here(), "data-updated/species-information/species_list_for_sdms.csv"), row.names = F)
```
