==> src/data-wrangling/04_generate_grids.Rmd <==
---
title: "Generate 1x1 and 5x5 Grids for Model Inputs"
output: html_notebook
---

This script takes the environmental, GFW, and SDM data and creates bins at 1 or 5 degree resolution to match with RFMO reported data. 

```{r}
# Load packages
library(tidyverse)
library(readxl)
library(googledrive)
library(rfishbase) # just for nei species
library(here)

###
# Load data
###

# Species lists for sdms 
species_list <- read.csv(file.path(here::here(), "data-updated/species-information/species_list_for_sdms.csv"))

### 
# Set Values
###
spatial_scale <- 5 # adjust depending on the scale we're aiming for (5x5 cell or 1x1 cell)
```

# IUCN SDMs
```{r}
# Load, clean, and filter SDM file names
sdm_list <- list.files(file.path(here::here(), "data-updated/iucn-sdm-data/intermediates"))

# Create master database with sdm info
all_sdm <- NULL
for(file in sdm_list) {
  
  temp_sdm <- read.csv(file.path(here::here(), "data-updated/iucn-sdm-data/intermediates", file))

  colnames(temp_sdm)[4] <- "probability"

  temp_sdm_save <- NULL

  for(shift_lat_n in 1:2) {
    for(shift_lon_n in 1:2) {
      temp_shift <- temp_sdm
      if(shift_lat_n == 1) {
        temp_shift <- temp_shift %>%
          mutate(lat_bin = as.numeric(paste(cut(y, breaks = seq(-97.5, 92.5,spatial_scale),
                                                labels = seq(-97.5 + spatial_scale, 92.5, spatial_scale)))),
                 lat_bin = case_when(lat_bin == 92.5 ~ 87.5,
                                     lat_bin == -97.5 + spatial_scale ~ -87.5,
                                     TRUE ~ lat_bin))
      }
      if(shift_lat_n == 2) {
        temp_shift <- temp_shift %>%
          mutate(lat_bin = as.numeric(paste(cut(y, breaks = seq(-95, 90, spatial_scale),
                                        labels = seq(-95 + spatial_scale, 90, spatial_scale)))))
      }
      if(shift_lon_n == 1) {
        temp_shift <- temp_shift %>%
          mutate(lon_bin = as.numeric(paste(cut(x, breaks = seq(-187.5, 182.5, spatial_scale),
                                        labels = seq(-187.5 + spatial_scale, 182.5, spatial_scale)))),
                 lon_bin = case_when(lon_bin == 182.5 ~ 177.5,
                                     lon_bin == -187.5 + spatial_scale ~ -177.5,
                                     TRUE ~ lon_bin))
      }
      if(shift_lon_n == 2){
        temp_shift <- temp_shift %>%
          mutate(lon_bin = as.numeric(paste(cut(x, breaks = seq(-185, 180, spatial_scale),
                                                labels = seq(-185 + spatial_scale, 180, spatial_scale)))))
        }
      temp_sdm_save <- temp_sdm_save %>%
        bind_rows(temp_shift)
    }
  }

  temp_sdm_save <- temp_sdm_save %>%
    mutate(sp_name = toupper(sp_name),
           probability = ifelse(is.na(probability), 0, probability)) %>%
    group_by(lat_bin, lon_bin, sp_name) %>%
    summarise(mean_sdm = mean(probability)) %>%
    ungroup() 

  if(! (temp_sdm_save$sp_name[1] %in% unique(all_sdm$sp_name))) {
    all_sdm <- all_sdm %>% bind_rows(temp_sdm_save)
  }
}

# keep only important
all_sdm <- all_sdm %>%
  rename(latitude = lat_bin,
         longitude = lon_bin,
         species_sciname = sp_name,
         sdm = mean_sdm)

# Get list of species names and file names... for nei grouping
sdm_files_names <- NULL
for(file in sdm_list) {
  temp_sdm <- read_csv(file.path(here::here(), "data-updated/iucn-sdm-data/intermediates", file), n_max = 1)
  sdm_files_names <- bind_rows(sdm_files_names, data.frame("file" = file, "species_sciname" = temp_sdm$sp_name))
}

# NEIs options
nei_sdms <- NULL
for(nei in unique(species_list$species_sciname)) {
  temp_group <- species_list %>% filter(species_sciname == nei)
  temp_sdm_list <- sdm_files_names[sdm_files_names$species_sciname %in% c(temp_group$value),"file"]

  nei_sdms_sub <- NULL

  if(length(temp_sdm_list) != 0) {
    for(file in temp_sdm_list) {

    temp_sdm <- read_csv(file.path(here::here(),"data-updated/iucn-sdm-data/intermediates", file))

    colnames(temp_sdm)[4] <- "probability"

    temp_sdm_save <- NULL

    for(shift_lat_n in 1:2) {
      for(shift_lon_n in 1:2) {
        temp_shift <- temp_sdm
        if(shift_lat_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(y, breaks = seq(-97.5, 92.5, spatial_scale),
                                                  labels = seq(-97.5 + spatial_scale, 92.5, spatial_scale)))),
                   lat_bin = case_when(lat_bin == 92.5 ~ 87.5,
                                       lat_bin == -97.5 + spatial_scale ~ -87.5,
                                       TRUE ~ lat_bin))
        }
        if(shift_lat_n == 2) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(y, breaks = seq(-95, 90, spatial_scale),
                                          labels = seq(-95 + spatial_scale, 90, spatial_scale)))))
        }
        if(shift_lon_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(x, breaks = seq(-187.5, 182.5, spatial_scale),
                                          labels = seq(-187.5 + spatial_scale, 182.5, spatial_scale)))),
                   lon_bin = case_when(lon_bin == 182.5 ~ 177.5,
                                       lon_bin == -187.5 + spatial_scale ~ -177.5,
                                       TRUE ~ lon_bin))
        }
        if(shift_lon_n == 2){
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(x, breaks = seq(-185, 180, spatial_scale),
                                                  labels = seq(-185 + spatial_scale, 180, spatial_scale)))))
          }
        temp_sdm_save <- temp_sdm_save %>%
          bind_rows(temp_shift)
      }
  }

  temp_sdm_save <- temp_sdm_save %>%
    mutate(sp_name = toupper(sp_name),
           probability = ifelse(is.na(probability), 0, probability)) %>%
    group_by(lat_bin, lon_bin, sp_name) %>%
    summarise(mean_sdm = mean(probability)) %>%
    ungroup()

  nei_sdms_sub <- nei_sdms_sub %>% bind_rows(temp_sdm_save)

  }
    nei_sdms_sub <- nei_sdms_sub %>%
      rename(probability = mean_sdm) %>%
      mutate(sp_name = nei, 
             probability = ifelse(is.na(probability), 0, probability)) %>%
      group_by(lat_bin, lon_bin, sp_name) %>%
      summarise(mean_sdm = mean(probability)) %>%
      ungroup()

    nei_sdms <- nei_sdms %>%
      bind_rows(nei_sdms_sub)
  }
}

nei_sdms <- nei_sdms %>%
  rename(latitude = lat_bin,
         longitude = lon_bin,
         species_sciname = sp_name,
         sdm = mean_sdm)

all_sdm <- all_sdm %>%
  bind_rows(nei_sdms) %>% 
  distinct_all()

write.csv(all_sdm, file.path(here::here(), paste0("data-updated/iucn-sdm-data/outputs/global_combined_iucn_sdm_",spatial_scale,"x",spatial_scale,".csv")),
          row.names = F)
```

# Sea Surface Temperature
```{r}
file_list <- list.files(file.path(here::here(), "data-updated/sea-surface-temperature/intermediates"),
                        pattern = "csv")

sst_data <- NULL

for(file in file_list) {
  temp <- read.csv(file.path(here::here(), "data-updated/sea-surface-temperature/intermediates", file))
  temp_save <- NULL

 for(shift_lat_n in 1:2) {
      for(shift_lon_n in 1:2) {
        temp_shift <- temp
        if(shift_lat_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(lat, breaks = seq(-97.5, 92.5, spatial_scale),
                                                  labels = seq(-97.5 + spatial_scale, 92.5, spatial_scale)))),
                   lat_bin = case_when(lat_bin == 92.5 ~ 87.5,
                                       lat_bin == -97.5 + spatial_scale ~ -87.5,
                                       TRUE ~ lat_bin))
        }
        if(shift_lat_n == 2) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(lat, breaks = seq(-95, 90, spatial_scale),
                                          labels = seq(-95 + spatial_scale, 90, spatial_scale)))))
        }
        if(shift_lon_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(lon, breaks = seq(-187.5, 182.5, spatial_scale),
                                          labels = seq(-187.5 + spatial_scale, 182.5, spatial_scale)))),
                   lon_bin = case_when(lon_bin == 182.5 ~ 177.5,
                                       lon_bin == -187.5 + spatial_scale ~ -177.5,
                                       TRUE ~ lon_bin))
        }
        if(shift_lon_n == 2){
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(lon, breaks = seq(-185, 180, spatial_scale),
                                                  labels = seq(-185 + spatial_scale, 180, spatial_scale)))))
          }
        temp_save <- temp_save %>%
          bind_rows(temp_shift)
      }
  }

  temp_save <- temp_save %>%
    filter(!is.na(sst)) %>%
    mutate(year = lubridate::year(lubridate::as_date(time))) %>%
    group_by(year, lon_bin, lat_bin) %>%
    summarise(mean_sst = mean(sst, na.rm = T),
              median_sst = median(sst, na.rm = T),
              min_sst = min(sst, na.rm = T),
              max_sst = max(sst, na.rm = T),
              sd_sst = sd(sst, na.rm = T),
              se_sst = sd_sst/(sqrt(n())),
              cv_sst = sd_sst/mean_sst*100) %>%
     ungroup() %>% 
    rename(longitude = lon_bin, 
           latitude = lat_bin)

  sst_data <- sst_data %>%
    bind_rows(temp_save) %>% 
    distinct_all()

}

write.csv(sst_data, file.path(here::here(), paste0("data-updated/sea-surface-temperature/outputs/binned_global_sst_",spatial_scale,"x",spatial_scale,".csv")),
          row.names = F)
```

# Chlorophyll-A
```{r}
file_list <- list.files(file.path(here::here(), "data-updated/chlorophyll-a/intermediates"),
                        pattern = "csv")

chla_data <- NULL

for(file in file_list) {
  temp <- read.csv(file.path(here::here(), "data-updated/chlorophyll-a/intermediates",
                             file))
  temp_save <- NULL

  for(shift_lat_n in 1:2) {
    for(shift_lon_n in 1:2) {
      temp_shift <- temp
      if(shift_lat_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(lat, breaks = seq(-97.5, 92.5, spatial_scale),
                                                  labels = seq(-97.5 + spatial_scale, 92.5, spatial_scale)))),
                   lat_bin = case_when(lat_bin == 92.5 ~ 87.5,
                                       lat_bin == -97.5 + spatial_scale ~ -87.5,
                                       TRUE ~ lat_bin))
        }
        if(shift_lat_n == 2) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(lat, breaks = seq(-95, 90, spatial_scale),
                                          labels = seq(-95 + spatial_scale, 90, spatial_scale)))))
        }
        if(shift_lon_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(lon, breaks = seq(-187.5, 182.5, spatial_scale),
                                          labels = seq(-187.5 + spatial_scale, 182.5, spatial_scale)))),
                   lon_bin = case_when(lon_bin == 182.5 ~ 177.5,
                                       lon_bin == -187.5 + spatial_scale ~ -177.5,
                                       TRUE ~ lon_bin))
        }
        if(shift_lon_n == 2){
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(lon, breaks = seq(-185, 180, spatial_scale),
                                                  labels = seq(-185 + spatial_scale, 180, spatial_scale)))))
          }
      temp_save <- temp_save %>%
        bind_rows(temp_shift)
    }
  }

  temp_save <- temp_save %>%
    filter(!is.na(chlor_a)) %>%
    rename(chla = chlor_a) %>%
    mutate(year = lubridate::year(lubridate::as_date(time))) %>%
    group_by(year, lon_bin, lat_bin) %>%
    summarise(mean_chla = mean(chla, na.rm = T),
              median_chla = median(chla, na.rm = T),
              min_chla = min(chla, na.rm = T),
              max_chla = max(chla, na.rm = T),
              sd_chla = sd(chla, na.rm = T),
              se_chla = sd_chla/(sqrt(n())),
              cv_chla = sd_chla/mean_chla*100) %>%
    ungroup() %>% 
    rename(longitude = lon_bin, 
           latitude = lat_bin)

  chla_data <- chla_data %>%
    bind_rows(temp_save) %>% 
    distinct_all()

}

write.csv(chla_data, file.path(here::here(), paste0("data-updated/chlorophyll-a/outputs/binned_global_chla_",spatial_scale,"x",spatial_scale,".csv")),
          row.names = F)
```

# Sea Surface Height
```{r}
# grab data first
# List files
# files <- list.files(file.path(here::here(), "data-updated", "sea-surface-height", "inputs"))
# 
# for(file in files) { 
#   tmp_raster <- brick(file.path(here::here(), "data-updated", "sea-surface-height", "inputs", file), varname = "SLA")
#   
#   tmp_df <- as.data.frame(tmp_raster, xy = TRUE) %>% 
#     rename(layer = colnames(.)[3]) %>% 
#     rename(longitude = y, 
#            latitude = x) %>% 
#     mutate(longitude = ifelse(longitude > 180, longitude - 360, longitude),  
#            date = as.Date("1985-01-01") + as.numeric(gsub("X", "", names(tmp_raster)))) %>% 
#     filter(!is.na(layer))
#   
#   write.csv(tmp_df, file.path(here::here(), "data-updated", "sea-surface-height", "intermediates", 
#                               gsub("[.]nc", ".csv", file)), row.names = F)
#   }
# 
# # Take means across years... 
# for(year in 2012:2019) { 
#   files <- list.files(file.path(here::here(), "data-updated", "sea-surface-height", "intermediates"), 
#                       pattern = paste0("ssh_grids_v1812_", year), full.names = T)
#   
#   all_data <- NULL
#   
#   for(file in files) { 
#     temp <- read.csv(file)
#     
#     all_data <- bind_rows(all_data, temp)
#     
#   }
#   
#   all_data <- all_data %>% 
#     group_by(latitude, longitude) %>% 
#     summarise(mean_ssh = mean(layer, na.rm = T)) %>% 
#     ungroup() %>% 
#     mutate(year = year)
#   
#   write.csv(all_data, file.path(here::here(), "data-updated", "sea-surface-height", "intermediates", 
#                                 paste0("mean_ssh_", year, ".csv")), 
#             row.names = F)
#   
#   print(year)
# } 

# Now we can generate the grids
file_list <- list.files(file.path(here::here(), "data-updated/sea-surface-height/intermediates"),
                        pattern = "mean_ssh")

ssh_data <- NULL

for(file in file_list) {
  temp <- read.csv(file.path(here::here(), "data-updated/sea-surface-height/intermediates",
                             file))
  temp_save <- NULL

 for(shift_lat_n in 1:2) {
      for(shift_lon_n in 1:2) {
        temp_shift <- temp
        if(shift_lat_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(latitude, breaks = seq(-97.5, 92.5, spatial_scale),
                                                  labels = seq(-97.5 + spatial_scale, 92.5, spatial_scale)))),
                   lat_bin = case_when(lat_bin == 92.5 ~ 87.5,
                                       lat_bin == -97.5 + spatial_scale ~ -87.5,
                                       TRUE ~ lat_bin))
        }
        if(shift_lat_n == 2) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(latitude, breaks = seq(-95, 90, spatial_scale),
                                          labels = seq(-95 + spatial_scale, 90, spatial_scale)))))
        }
        if(shift_lon_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(longitude, breaks = seq(-187.5, 182.5, spatial_scale),
                                          labels = seq(-187.5 + spatial_scale, 182.5, spatial_scale)))),
                   lon_bin = case_when(lon_bin == 182.5 ~ 177.5,
                                       lon_bin == -187.5 + spatial_scale ~ -177.5,
                                       TRUE ~ lon_bin))
        }
        if(shift_lon_n == 2){
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(longitude, breaks = seq(-185, 180, spatial_scale),
                                                  labels = seq(-185 + spatial_scale, 180, spatial_scale)))))
          }
        temp_save <- temp_save %>%
          bind_rows(temp_shift)
      }
  }

  temp_save <- temp_save %>%
    rename(ssh = mean_ssh) %>%
    filter(!is.na(ssh)) %>%
    group_by(year, lon_bin, lat_bin) %>%
    summarise(mean_ssh = mean(ssh, na.rm = T),
              median_ssh = median(ssh, na.rm = T),
              min_ssh = min(ssh, na.rm = T),
              max_ssh = max(ssh, na.rm = T),
              sd_ssh = sd(ssh, na.rm = T),
              se_ssh = sd_ssh/(sqrt(n())),
              cv_ssh = sd_ssh/mean_ssh*100) %>%
     ungroup() %>%
    rename(longitude = lon_bin,
           latitude = lat_bin)

  ssh_data <- ssh_data %>%
    bind_rows(temp_save) %>% 
    distinct_all()

}

write.csv(ssh_data, file.path(here::here(), paste0("data-updated/sea-surface-height/outputs/binned_global_ssh_",spatial_scale,"x",spatial_scale,".csv")),
          row.names = F)
```

# Global Fishing Watch
```{r}
gfw_data <- vroom::vroom(file.path(here::here(), "data-updated/global-fishing-watch/inputs/gfw_effort_ll_ps_1x1_2012_2021.csv"))

tonnage_groups <- read.csv(file.path(here::here(), "data-updated/global-fishing-watch/inputs/gfw_tonnage_groups_v20220803.csv"))

# simplify for easier use
gfw <- gfw_data %>%
  filter(rfmo %in% c("WCPFC", "IOTC", "ICCAT", "IATTC")) %>%
  mutate(gear_group = case_when(grepl("longline", gear) ~ "longline",
                                grepl("purse", gear) ~ "purse seine",
                                TRUE ~ "other")) %>%
  filter(gear_group != "other") %>%
  group_by(year, lat_bin, lon_bin, rfmo, gear_group) %>%
  summarise(total_fishing_kwh = sum(total_fishing_kwh, na.rm = T)) %>%
  ungroup()

# add different shift in case needed
gfw_shift <- NULL

for(shift_lat_n in 1:2) {
  for(shift_lon_n in 1:2) {
    temp_shift <- gfw
      if(shift_lat_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(lat_bin, breaks = seq(-97.5, 92.5, spatial_scale),
                                                  labels = seq(-97.5 + spatial_scale, 92.5, spatial_scale)))),
                   lat_bin = case_when(lat_bin == 92.5 ~ 87.5,
                                       lat_bin == -97.5 + spatial_scale ~ -87.5,
                                       TRUE ~ lat_bin))
        }
        if(shift_lat_n == 2) {
          temp_shift <- temp_shift %>%
            mutate(lat_bin = as.numeric(paste(cut(lat_bin, breaks = seq(-95, 90, spatial_scale),
                                          labels = seq(-95 + spatial_scale, 90, spatial_scale)))))
        }
        if(shift_lon_n == 1) {
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(lon_bin, breaks = seq(-187.5, 182.5, spatial_scale),
                                          labels = seq(-187.5 + spatial_scale, 182.5, spatial_scale)))),
                   lon_bin = case_when(lon_bin == 182.5 ~ 177.5,
                                       lon_bin == -187.5 + spatial_scale ~ -177.5,
                                       TRUE ~ lon_bin))
        }
        if(shift_lon_n == 2){
          temp_shift <- temp_shift %>%
            mutate(lon_bin = as.numeric(paste(cut(lon_bin, breaks = seq(-185, 180, spatial_scale),
                                                  labels = seq(-185 + spatial_scale, 180, spatial_scale)))))
          }
   gfw_shift <- gfw_shift %>%
     bind_rows(temp_shift)
 }
}


gfw_shift <- gfw_shift %>%
  group_by(year, lat_bin, lon_bin, rfmo, gear_group) %>%
  summarise(total_fishing_kwh = sum(total_fishing_kwh, na.rm = T)) %>%
  ungroup() %>%
  rename(latitude = lat_bin,
         longitude = lon_bin) %>%
  distinct_all()
# save
write.csv(gfw_shift,  file.path(here::here(), paste0("data-updated/global-fishing-watch/outputs/binned_global_gfw_",spatial_scale,"x",spatial_scale,".csv")), row.names = F)
```
