==> src/data-wrangling/05_all_rfmo_model_data_cleaning_collation.Rmd <==
---
title: "Data Cleaning for ML - All RFMO data"
output: html_notebook
---

This notebook combines RFMO catch data with SDM, SST, chl-a, and GFW data for all RFMOs using 5x5 and 1x1 degree grids.

Data that are reported in 1x1 cells are left as is. Data that are reported in 5x5 cells are converted to 1x1 cells using catch = catch/25. 

Data that are reported as counts are left as is. Data that are reported as metric tonnes are converted to count using established length-weight relationships and common lengths from Fishbase to calculate a common weight (mt). Reported catch (mt) is divided by the common weight (mt) to get count. 

```{r message=FALSE, warning=FALSE, include=FALSE}
# Load Libraries
library(tidyverse)
library(readxl)
library(googledrive)
library(sf)
library(rgdal)
library(here)
```

```{r include=FALSE}
# Load Data
all_data <- read.csv(file.path(here::here(), "data-updated/rfmo-data/outputs/all_data.csv"))

# # Species grouping
# species_groups <- read_xlsx(file.path(here::here(), "data-updated/species-information/species_groups.xlsx"), sheet = "species_list") %>% 
#   filter(group %in% c("tunas", "tuna-like species", "sharks and rays")) %>% 
#   mutate(scientific_name = str_to_upper(scientific_name))

# SDM data
all_sdm_1x1 <- read.csv(file.path(here::here(), "data-updated/iucn-sdm-data/outputs/global_combined_iucn_sdm_1x1.csv"))
all_sdm_5x5 <- read.csv(file.path(here::here(), "data-updated/iucn-sdm-data/outputs/global_combined_iucn_sdm_5x5.csv"))

# SST data
sst_data_1x1 <- read.csv(file.path(here::here(), "data-updated/sea-surface-temperature/outputs/binned_global_sst_1x1.csv"))
sst_data_5x5 <- read.csv(file.path(here::here(), "data-updated/sea-surface-temperature/outputs/binned_global_sst_5x5.csv"))

# SSH data
ssh_data_1x1 <- read.csv(file.path(here::here(), "data-updated/sea-surface-height/outputs/binned_global_ssh_1x1.csv"))
ssh_data_5x5 <- read.csv(file.path(here::here(), "data-updated/sea-surface-height/outputs/binned_global_ssh_5x5.csv"))

# Chlorophyll-A data
chla_data_1x1 <- read.csv(file.path(here::here(), "data-updated/chlorophyll-a/outputs/binned_global_chla_1x1.csv"))
chla_data_5x5 <- read.csv(file.path(here::here(), "data-updated/chlorophyll-a/outputs/binned_global_chla_5x5.csv"))

# GFW data
gfw_data_1x1 <- read.csv(file.path(here::here(), "data-updated/global-fishing-watch/outputs/binned_global_gfw_1x1.csv"))
gfw_data_5x5 <- read.csv(file.path(here::here(), "data-updated/global-fishing-watch/outputs/binned_global_gfw_5x5.csv"))

# Ex-vessel price data
prices <- read.csv(file.path(here::here(), "data-updated/ex-vessel-prices/exvessel_price_database_1976_2019.csv")) %>% 
  distinct_all()

# add the min/max years for each year - based on the most recent year that all data sources
# for an RFMO have data
min_yr = 2012 
max_yr = data.frame("year" = c(2020, 2020, 2020, 2020), 
                    "rfmo" = c("IOTC", "IATTC", "ICCAT", "WCPFC")) 
  #all RFMOS have this as their max year for LL 
```


```{r}
# Fix the prices data
prices <- prices %>% 
  mutate(ASFIS_species = str_to_upper(ASFIS_species), 
         scientific_name = str_to_upper(scientific_name), 
         ASFIS_species = case_when(ASFIS_species == "INDO-PACIF. BOTTLENOSE DOLPHIN" ~  "INDO-PACIFIC BOTTLENOSE DOLPHIN", 
                                   ASFIS_species == "SHORTFIN MAKO" ~ "SHORTFIN MAKO SHARK", 
                                   ASFIS_species == "ALBACORE" ~ "ALBACORE TUNA", 
                                   TRUE ~ ASFIS_species), 
         ISSCAAP_group = case_when(ISSCAAP_group == "Sharks, rays, chimaeras" ~ "sharks and rays", 
                                   ISSCAAP_group == "Miscellaneous aquatic mammals" ~ "marine mammals", 
                                   ISSCAAP_group == "Turtles" ~ "turtles", 
                                   TRUE ~ paste(ISSCAAP_group))) %>% 
  dplyr::rename(year = Year, 
         species_commonname = ASFIS_species, 
         species_group = ISSCAAP_group)

###
# median prices by species group
###
spp_prices <- prices %>% 
  mutate(scientific_name = gsub(" SPP", "", scientific_name)) %>%
  filter(scientific_name %in% unique(all_data$species_sciname)) %>% 
  group_by(year) %>% 
  summarise(median_price_group = median(exvessel, na.rm = T)) %>% 
  ungroup() %>% 
  filter(year >= min_yr) 

# adjust for 2020
spp_prices <- spp_prices %>% 
  bind_rows(spp_prices %>% 
              filter(year == 2019) %>% 
              mutate(year = 2020, 
                     median_price_group = median_price_group * 1.012)) %>% 
  bind_rows(spp_prices %>% 
              filter(year == 2019) %>% 
              mutate(year = 2021, 
                     median_price_group = median_price_group * 1.06))

###
# target prices for individuals
###
spp_prices_ind <- prices %>% 
  mutate(scientific_name = gsub(" SPP", "", scientific_name)) %>% 
  bind_rows(prices %>% filter(grepl("SPHYRNA", scientific_name)) %>% 
              group_by(year) %>% 
              summarise(exvessel = median(exvessel, na.rm = T)) %>% 
              ungroup() %>% 
              mutate(scientific_name = "SPHYRNA")) %>% 
  filter(scientific_name %in% unique(all_data$species_sciname)) %>% 
  group_by(year, scientific_name) %>% 
  summarise(median_price_species = median(exvessel, na.rm = T)) %>% 
  ungroup() %>% 
  filter(year >= min_yr) 

# Adjust for 2020
spp_prices_ind <- spp_prices_ind  %>% 
  bind_rows(spp_prices_ind %>% 
              filter(year == 2019) %>% 
              mutate(year = 2020, 
                     median_price_species = median_price_species * 1.012)) %>% 
  bind_rows(spp_prices_ind %>% 
              filter(year == 2019) %>% 
              mutate(year = 2021, 
                     median_price_species = median_price_species * 1.06))

# Add species that aren't there
spp_prices_ind <- spp_prices_ind  %>% 
  bind_rows(spp_prices %>% 
              mutate(scientific_name = "SHARKS NEI") %>% 
              dplyr::rename(median_price_species = median_price_group) %>% 
              select(year, scientific_name, median_price_species)) %>% 
    bind_rows(spp_prices %>% 
              mutate(scientific_name = "RHINCODON TYPUS") %>% # don't have any data for whaleshark so we will just use all
              dplyr::rename(median_price_species = median_price_group) %>% 
              select(year, scientific_name, median_price_species))
```

```{r message=FALSE, warning=FALSE}
for(rfmos in unique(all_data$rfmo)) { 
  
  temp <- all_data %>% 
    filter(rfmo == rfmos &
             spatial_notes %in% c("center of 5x5 cell", "center of 1x1 cell") &
           gear_group == "longline" & #effort_units == "hooks" &
             catch_units %in% c("metric tonnes", "count") & 
           was_generated == "no" &
           year <= max_yr$year[which(max_yr$rfmo == rfmos)])
  
  if(rfmos == "IOTC") { 
    temp <- temp %>% 
      filter(sources != "IOTC-2020-WPEB16-DATA12_CE")
  }
  
  if(rfmos == "WCPFC") { 
    temp <- temp %>% 
    select(rfmo, year, country, latitude, longitude, gear_group, time_period, effort, effort_units, spatial_notes, species_group, 
           species_commonname, species_sciname, catch, catch_units) %>% 
    distinct_all() %>% 
    group_by(rfmo, year, country, latitude, longitude, gear_group, time_period, effort_units,
             spatial_notes, species_group, species_commonname, species_sciname, catch_units) %>% 
    summarise(catch = sum(catch, na.rm = T), 
              effort = sum(effort, na.rm = T)) %>% 
    ungroup() 
  } else {
    temp <- temp %>%
      select(rfmo, year, country, latitude, longitude, gear_group, time_period, effort, effort_units, spatial_notes, species_group, 
             species_commonname, species_sciname, catch, catch_units) %>% 
      distinct_all() %>% 
      group_by(rfmo, year, country, latitude, longitude, gear_group, time_period, effort_units,
               spatial_notes, species_group, species_commonname, species_sciname, catch_units) %>% 
      summarise(catch = sum(catch, na.rm = T), 
                effort = sum(effort, na.rm = T)) %>% 
      ungroup() %>% 
      mutate(was_generated = "yes", 
             aggregation_period = time_period, 
             time_period = "yearly") 
    
    if("monthly" %in% temp$aggregation_period) { 
      temp <- temp %>% 
        filter(aggregation_period == "monthly")
    } else { 
      temp <- temp %>% 
        filter(aggregation_period == "quarterly")
    }
  } 
  
  short_data <- temp

# Scale to lowest possible resolution
short_data <- short_data %>% 
  mutate(species_sciname = ifelse(is.na(species_sciname), species_commonname, species_sciname)) %>%
  group_by(rfmo, country, year, latitude, longitude, gear_group, effort_units, species_commonname, spatial_notes,species_group, 
           species_sciname, catch_units) %>% 
  summarise(catch = sum(catch, na.rm = T), 
            effort = sum(effort, na.rm = T)) %>% 
  ungroup() 

# short_data <- short_data %>% 
#   left_join(species_groups %>% 
#               dplyr::rename(species_group = group, 
#                      species_commonname = english_name) %>% 
#               mutate(species_commonname = str_to_upper(species_commonname)) %>% 
#               select(-species_commonname),
#             by = c("species_sciname" = "scientific_name")) %>% 
#   select(rfmo, year, country, latitude, longitude, gear_group, effort, effort_units, species_group, spatial_notes,
#          species_commonname, species_sciname, catch, catch_units) %>%
#   mutate(species_sciname = case_when(species_sciname == "TETRAPTURUS AUDAX" ~ "KAJIKIA AUDAX",
#                                      species_sciname == "MAKAIRA INDICA" ~ "ISTIOMPAX INDICA", 
#                                      species_sciname == "TETRAPTURUS ALBIDUS" ~ "KAJIKIA ALBIDA",
#                                      species_sciname == "TETRAPTURUS PFLUEGERI" ~ "TETRAPTURUS PFLUEGERI", 
#                                      TRUE ~ species_sciname))

# how much data contributed to "nei"s
# short_data <- short_data %>% 
#   mutate(species_group = case_when(is.na(species_group) & 
#                                      grepl("NEI|OTHER", species_commonname) & 
#                                      grepl("SHARK", species_commonname) ~ "sharks and rays", 
#                                    species_commonname == "SILKY OR BLACKTIP SHARK" ~ "sharks and rays",
#                                    is.na(species_group) & 
#                                      grepl("NEI|OTHER", species_commonname) & 
#                                      grepl("MARLIN|BILLFISH|TUNA-LIKE", species_commonname) ~ "tuna-like species", 
#                                    species_commonname == "FRIGATE AND BULLET TUNAS" ~ "tunas",
#                                    is.na(species_group) & 
#                                      grepl("NEI|OTHER", species_commonname) & 
#                                      grepl("TUNA", species_commonname) ~ "tunas",
#                                    TRUE ~ species_group))

# Now standardize nei species if can...
short_data <- short_data %>%
  mutate(species_resolution = case_when(grepl("NEI|OTHER|SPP", species_commonname) ~ "NEI",
                                        grepl("NEI|OTHER|SPP", species_sciname) ~ "NEI",
                                        species_commonname == "SILKY OR BLACKTIP SHARK" ~ "NEI",
                                        species_commonname == "FRIGATE AND BULLET TUNAS" ~ "NEI",
                                        !grepl(" ", species_sciname) ~ "NEI",
                                        TRUE ~ "SPECIES SPECIFIC"),
         species_commonname = case_when(species_commonname == "OTHER SHARKS" ~ "SHARKS NEI",
                                        species_commonname == "OTHER TUNA" ~ "TUNAS NEI",
                                        TRUE ~ species_commonname),
         species_sciname = case_when(species_commonname == "TUNAS NEI" ~ "THUNNUS",
                                     species_commonname == "SHARKS NEI" ~ "SHARKS NEI", 
                                     is.na(species_sciname) ~ species_commonname,
                                     TRUE ~ gsub(" SPP|[.]", "", species_sciname))) 

# Calculate effort for model runs
target_effort_pivot <- short_data %>% 
  filter(species_group %in% c("tunas", "tuna-like species")) %>% 
  filter(effort_units == "hooks") %>% 
  dplyr::select(rfmo, year, country, latitude, longitude, gear_group, effort_units, effort, spatial_notes) %>% 
  distinct_all() %>% 
  group_by(rfmo, year, country, latitude, longitude, gear_group, effort_units, spatial_notes) %>% 
  summarise(effort = round(sum(effort, na.rm = T))) %>% 
  ungroup()

target_effort_pivot <- target_effort_pivot %>% 
  group_by(rfmo, year, latitude, longitude, gear_group, effort_units, spatial_notes) %>% 
  summarise(target_effort = round(sum(effort, na.rm = T))) %>% 
  ungroup() %>%
  full_join(target_effort_pivot  %>%
               pivot_wider(names_from = country, 
                           values_from = effort, 
                           values_fn = sum, 
                           values_fill = 0), 
            by = c("year","rfmo", "latitude", "longitude", "gear_group", "effort_units", "spatial_notes")) 

colnames(target_effort_pivot)[9:ncol(target_effort_pivot)] <- paste0("target_effort_",colnames(target_effort_pivot)[9:ncol(target_effort_pivot)], "_longline")

# Effort reported with bycatch spp
bycatch_effort_pivot <- short_data %>% 
  filter(species_group %in% c("sharks and rays")) %>% 
  filter(effort_units == "hooks") %>% 
  dplyr::select(rfmo, year, country, latitude, longitude, gear_group, effort_units, effort, spatial_notes) %>% 
  distinct_all() %>% 
  group_by(rfmo, year, country, latitude, longitude, gear_group, effort_units, spatial_notes) %>% 
  summarise(effort = round(sum(effort, na.rm = T))) %>% 
  ungroup()

bycatch_effort_pivot <- bycatch_effort_pivot %>% 
  group_by(rfmo, year, latitude, longitude, gear_group, effort_units, spatial_notes) %>% 
  summarise(bycatch_total_effort = round(sum(effort, na.rm = T))) %>% 
  ungroup() %>%
  full_join(bycatch_effort_pivot  %>%
               pivot_wider(names_from = country, 
                           values_from = effort, 
                           values_fn = sum, 
                           values_fill = 0), 
            by = c("year","rfmo", "latitude", "longitude", "gear_group", "effort_units", "spatial_notes")) 

colnames(bycatch_effort_pivot)[9:ncol(bycatch_effort_pivot)] <- paste0("bycatch_total_effort_",colnames(bycatch_effort_pivot)[9:ncol(bycatch_effort_pivot)], "_longline")

# quick gather to include tuna species as well
short_data_inc_tuna <- short_data

short_data_inc_tuna <- short_data_inc_tuna %>% 
  group_by(rfmo, year, latitude, longitude, gear_group, effort_units, species_group, spatial_notes, species_commonname, species_sciname, catch_units, species_resolution) %>% 
  summarise(catch = sum(catch, na.rm = T)) %>% 
  ungroup() %>% 
  full_join(., target_effort_pivot %>% 
              mutate_at(vars(target_effort:colnames(target_effort_pivot)[length(colnames(target_effort_pivot))]), round)) %>%
  full_join(., bycatch_effort_pivot %>% 
              mutate_at(vars(bycatch_total_effort:colnames(bycatch_effort_pivot)[length(colnames(bycatch_effort_pivot))]), round)) %>%
  mutate(species_sciname = ifelse(is.na(species_sciname), "empty", species_sciname)) %>% 
  mutate_at(vars(target_effort:colnames(bycatch_effort_pivot)[length(colnames(bycatch_effort_pivot))]), replace_na, 0)

# Get nanes to add back later
names <- short_data_inc_tuna %>%
  dplyr::select(species_commonname, species_sciname, species_group, species_resolution, catch_units) %>%
  distinct_all() %>% 
  mutate(species_colnames = paste(species_sciname, catch_units, sep = "_"))

# Pad with zeros...
short_data_inc_tuna <- short_data_inc_tuna %>%
  dplyr::select(-species_commonname, - species_group, - species_resolution) %>%
  pivot_wider(names_from = c("species_sciname", "catch_units"),
              values_from = catch,
              values_fill = 0) %>%
  pivot_longer(cols = names$species_colnames,
               names_to = c("species_sciname",  "catch_units"), 
               names_sep = "_",
               values_to = "catch") %>% 
  relocate(species_sciname, .after = longitude) %>% 
  relocate(c('catch', 'catch_units'), .after = species_sciname) %>% 
  relocate(year, .after = rfmo) %>% 
  filter(species_sciname != "empty") %>% 
  left_join(., names %>% select(-species_colnames))

short_data_inc_tuna <- short_data_inc_tuna %>% 
  left_join(spp_prices, by = c("year")) %>% 
  left_join(spp_prices_ind, by = c("year", "species_sciname" = "scientific_name"))

short_data_inc_tuna <- short_data_inc_tuna %>% 
  distinct_all()

# Rescale to 1x1 degree 
locations_y <- short_data_inc_tuna %>%
  filter(spatial_notes == "center of 5x5 cell") %>%
  select(latitude) %>%
  distinct_all() %>%
  mutate(latitude_lag0 = latitude,
         latitude_lag1 = latitude-1,
         latitude_lag2 = latitude-2,
         latitude_plus1 = latitude+1,
         latitude_plus2 = latitude+2) %>%
  pivot_longer(latitude_lag0:latitude_plus2, values_to = "latitude_rescaled") %>%
  select(-name)

locations_x <- short_data_inc_tuna %>%
  filter(spatial_notes == "center of 5x5 cell") %>%
  select(longitude) %>%
  distinct_all() %>%
  mutate(longitude_lag0 = longitude,
         longitude_lag1 = longitude-1,
         longitude_lag2 = longitude-2,
         longitude_plus1 = longitude+1,
         longitude_plus2 = longitude+2) %>%
  pivot_longer(longitude_lag0:longitude_plus2, values_to = "longitude_rescaled") %>%
  select(-name)

# New dataset
short_data_inc_tuna_5x5 <- short_data_inc_tuna %>%
  filter(spatial_notes == "center of 5x5 cell") %>%
  left_join(locations_y) %>%
  left_join(locations_x) %>%
  mutate(longitude_rescaled = ifelse(is.na(longitude_rescaled), longitude, longitude_rescaled),
         latitude_rescaled = ifelse(is.na(latitude_rescaled), latitude, latitude_rescaled),
         catch = catch/25) %>%
  mutate_at(vars(target_effort:colnames(bycatch_effort_pivot)[length(colnames(bycatch_effort_pivot))]), function(x){ x/25}) %>% # rescale all to be at 1x1
  dplyr::rename(latitude_orig = latitude,
         longitude_orig = longitude,
         latitude = latitude_rescaled,
         longitude = longitude_rescaled) %>%
  mutate(spatial_notes = "center of 1x1 cell") %>%
  select(colnames(short_data_inc_tuna)) %>% 
  group_by_at(c(colnames(short_data_inc_tuna)[!grepl("target_effort|bycatch_total_effort|catch", colnames(short_data_inc_tuna))], "catch_units")) %>% 
  summarise_all(sum) %>% 
  ungroup()

short_data_inc_tuna <- short_data_inc_tuna %>%
  filter(spatial_notes != "center of 5x5 cell") %>%
  bind_rows(short_data_inc_tuna_5x5) %>%
  group_by_at(c(colnames(short_data_inc_tuna)[!grepl("target_effort|bycatch_total_effort|catch", colnames(short_data_inc_tuna))], "catch_units")) %>% 
  summarise_all(sum) %>% 
  ungroup()

short_data_inc_tuna <- short_data_inc_tuna %>% 
  distinct_all()

# Combine the data
# Using just the self-reported effort by hooks (so can only use catch by hooks)
combined_data <- short_data_inc_tuna %>%
  filter(effort_units == "hooks") %>% 
  mutate(year = as.integer(year)) %>% 
  filter(spatial_notes == "center of 1x1 cell") %>% 
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) # lower limit for GFW data

write.csv(combined_data,
          file.path(here::here(), 
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_1x1_tuna_hooks.csv")),
          row.names = FALSE)
### end break for including tuna

short_data <- short_data %>%
  filter(species_group == "sharks and rays") %>%
  group_by(rfmo, year, latitude, longitude, gear_group, effort_units, species_group, spatial_notes, species_commonname, species_sciname, catch_units, species_resolution) %>%
  summarise(catch = sum(catch, na.rm = T)) %>%
  ungroup() %>%
  full_join(., target_effort_pivot %>%
              mutate_at(vars(target_effort:colnames(target_effort_pivot)[length(colnames(target_effort_pivot))]), round)) %>%
  full_join(., bycatch_effort_pivot %>%
              mutate_at(vars(bycatch_total_effort:colnames(bycatch_effort_pivot)[length(colnames(bycatch_effort_pivot))]), round)) %>%
  mutate(species_sciname = ifelse(is.na(species_sciname), "empty", species_sciname)) %>%
  mutate_at(vars(target_effort:colnames(bycatch_effort_pivot)[length(colnames(bycatch_effort_pivot))]), replace_na, 0)

# Get nanes to add back later
names <- short_data %>%
  dplyr::select(species_commonname, species_sciname, species_group, species_resolution, catch_units) %>%
  distinct_all() %>%
  mutate(species_colnames = paste(species_sciname, catch_units, sep = "_"))

# Pad with zeros...
short_data <- short_data %>%
  dplyr::select(-species_commonname, - species_group, - species_resolution) %>%
  pivot_wider(names_from = c("species_sciname", "catch_units"),
              values_from = catch,
              values_fill = 0) %>%
  pivot_longer(cols = names$species_colnames,
               names_to = c("species_sciname",  "catch_units"),
               names_sep = "_",
               values_to = "catch") %>%
  relocate(species_sciname, .after = longitude) %>%
  relocate(c('catch', 'catch_units'), .after = species_sciname) %>%
  relocate(year, .after = rfmo) %>%
  filter(species_sciname != "empty") %>%
  left_join(., names %>% select(-species_colnames))

short_data <- short_data %>%
  left_join(spp_prices, by = c("year")) %>%
  left_join(spp_prices_ind, by = c("year", "species_sciname" = "scientific_name"))

short_data <- short_data %>%
  distinct_all()

# Remove data if 0s were filled for an entire year
short_data_remove <- short_data %>% 
  group_by(year, rfmo) %>% 
  summarise(sum_catch = sum(catch, na.rm = T)) %>% 
  ungroup() %>% 
  filter(sum_catch == 0)

short_data <- short_data %>% 
  filter(!year %in% short_data_remove$year)

# Combine the data
# Using just the self-reported effort by hooks (so can only use catch by hooks)
combined_data <- short_data %>%
  filter(effort_units == "hooks") %>%
  mutate(year = as.integer(year)) %>%
  filter(spatial_notes == "center of 5x5 cell" & catch_units == "count") %>%
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) %>% # lower limit for GFW data
  left_join(all_sdm_5x5 %>% distinct_all()) %>%
  #left_join(gfw_data_5x5 %>% distinct_all()) %>%
  left_join(sst_data_5x5 %>% distinct_all()) %>%
  left_join(chla_data_5x5 %>% distinct_all()) %>%
  left_join(ssh_data_5x5 %>% distinct_all())

write.csv(combined_data,
          file.path(here::here(),
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_5x5_count_hooks.csv")),
          row.names = FALSE)

# Using just GFW reported effort (so can use all catch reported, regardless of effort)
combined_data <- short_data %>%
  group_by(rfmo, year, latitude, longitude, gear_group, species_group, spatial_notes, species_commonname, species_sciname, catch_units, species_resolution) %>%
  summarise(catch = sum(catch, na.rm = T)) %>%
  ungroup() %>%
  mutate(year = as.integer(year)) %>%
  filter(spatial_notes == "center of 5x5 cell" & catch_units == "count") %>%
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) %>% # lower limit for GFW data
  left_join(all_sdm_5x5 %>% distinct_all()) %>%
  left_join(gfw_data_5x5 %>% distinct_all()) %>%
  left_join(sst_data_5x5 %>% distinct_all()) %>%
  left_join(chla_data_5x5 %>% distinct_all()) %>%
  left_join(ssh_data_5x5 %>% distinct_all())

write.csv(combined_data,
          file.path(here::here(),
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_5x5_count_kwh.csv")),
          row.names = FALSE)

if(length(unique(short_data$catch_units)) > 1) {
# Update mt to count
mt_count <- read.csv(file.path(here::here(), "data-updated/species-information/spp_weight_count_conversion.csv")) %>%
  mutate(scientific_name = str_to_upper(scientific_name))

# Add sharks nei to the mix
mt_count <- mt_count %>%
  bind_rows(data.frame(scientific_name = "SHARKS NEI",
                       weight_mt = mean(mt_count$weight_mt, na.rm = T))) %>%
  dplyr::select(scientific_name, weight_mt)

# Run conversions
short_data_mt <- short_data %>%
  filter(catch_units == "metric tonnes") %>%
  left_join(mt_count, by = c("species_sciname" = "scientific_name")) %>%
  mutate(catch = catch/weight_mt,
         catch_units = "count") %>%
  select(colnames(short_data))

# Add back to dataset
short_data2 <- short_data %>%
  filter(catch_units == "count") %>%
  bind_rows(short_data_mt) %>%
  group_by_at(colnames(short_data)[!colnames(short_data) == "catch"]) %>%
  summarise(catch = sum(catch, na.rm = T)) %>%
  ungroup()

short_data2 <- short_data2 %>%
  distinct_all()

# Combine the data

# Using just the self-reported effort by hooks (so can only use catch by hooks)
combined_data <- short_data2 %>%
  filter(effort_units == "hooks") %>%
  mutate(year = as.integer(year)) %>%
  filter(spatial_notes == "center of 5x5 cell" & catch_units == "count") %>%
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) %>% # lower limit for GFW data
  left_join(all_sdm_5x5 %>% distinct_all()) %>%
  #left_join(gfw_data_5x5 %>% distinct_all()) %>%
  left_join(sst_data_5x5 %>% distinct_all()) %>%
  left_join(chla_data_5x5 %>% distinct_all()) %>%
  left_join(ssh_data_5x5 %>% distinct_all())

write.csv(combined_data,
          file.path(here::here(),
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_5x5_mt_to_count_hooks.csv")),
          row.names = FALSE)

# Using just GFW reported effort (so can use all catch reported, regardless of effort)
combined_data <- short_data2 %>%
  group_by(rfmo, year, latitude, longitude, gear_group, species_group, spatial_notes, species_commonname, species_sciname, catch_units, species_resolution) %>%
  summarise(catch = sum(catch, na.rm = T)) %>%
  ungroup() %>%
  mutate(year = as.integer(year)) %>%
  filter(spatial_notes == "center of 5x5 cell" & catch_units == "count") %>%
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) %>% # lower limit for GFW data
  left_join(all_sdm_5x5 %>% distinct_all()) %>%
  left_join(gfw_data_5x5 %>% distinct_all()) %>%
  left_join(sst_data_5x5 %>% distinct_all()) %>%
  left_join(chla_data_5x5 %>% distinct_all())  %>%
  left_join(ssh_data_5x5 %>% distinct_all())

write.csv(combined_data,
          file.path(here::here(),
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_5x5_mt_to_count_kwh.csv")),
          row.names = FALSE)
}

# Rescale to 1x1 degree
locations_y <- short_data %>%
  filter(spatial_notes == "center of 5x5 cell") %>%
  select(latitude) %>%
  distinct_all() %>%
  mutate(latitude_lag0 = latitude,
         latitude_lag1 = latitude-1,
         latitude_lag2 = latitude-2,
         latitude_plus1 = latitude+1,
         latitude_plus2 = latitude+2) %>%
  pivot_longer(latitude_lag0:latitude_plus2, values_to = "latitude_rescaled") %>%
  select(-name)

locations_x <- short_data %>%
  filter(spatial_notes == "center of 5x5 cell") %>%
  select(longitude) %>%
  distinct_all() %>%
  mutate(longitude_lag0 = longitude,
         longitude_lag1 = longitude-1,
         longitude_lag2 = longitude-2,
         longitude_plus1 = longitude+1,
         longitude_plus2 = longitude+2) %>%
  pivot_longer(longitude_lag0:longitude_plus2, values_to = "longitude_rescaled") %>%
  select(-name)

# New dataset
short_data_5x5 <- short_data %>%
  filter(spatial_notes == "center of 5x5 cell") %>%
  left_join(locations_y) %>%
  left_join(locations_x) %>%
  mutate(longitude_rescaled = ifelse(is.na(longitude_rescaled), longitude, longitude_rescaled),
         latitude_rescaled = ifelse(is.na(latitude_rescaled), latitude, latitude_rescaled),
         catch = catch/25) %>%
  mutate_at(vars(target_effort:colnames(bycatch_effort_pivot)[length(colnames(bycatch_effort_pivot))]), function(x){ x/25}) %>% # rescale all to be at 1x1
  dplyr::rename(latitude_orig = latitude,
         longitude_orig = longitude,
         latitude = latitude_rescaled,
         longitude = longitude_rescaled) %>%
  mutate(spatial_notes = "center of 1x1 cell") %>%
  select(colnames(short_data)) %>%
  group_by_at(c(colnames(short_data)[!grepl("target_effort|bycatch_total_effort|catch", colnames(short_data))], "catch_units")) %>%
  summarise_all(sum) %>%
  ungroup()

short_data <- short_data %>%
  filter(spatial_notes != "center of 5x5 cell") %>%
  bind_rows(short_data_5x5) %>%
  group_by_at(c(colnames(short_data)[!grepl("target_effort|bycatch_total_effort|catch", colnames(short_data))], "catch_units")) %>%
  summarise_all(sum) %>%
  ungroup()

short_data <- short_data %>%
  distinct_all()

# Combine the data
# Using just the self-reported effort by hooks (so can only use catch by hooks)
combined_data <- short_data %>%
  filter(effort_units == "hooks") %>%
  mutate(year = as.integer(year)) %>%
  filter(spatial_notes == "center of 1x1 cell" & catch_units == "count") %>%
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) %>% # lower limit for GFW data
  left_join(all_sdm_1x1 %>% distinct_all()) %>%
  left_join(sst_data_1x1 %>% distinct_all()) %>%
  left_join(chla_data_1x1 %>% distinct_all()) %>%
  left_join(ssh_data_1x1 %>% distinct_all())

write.csv(combined_data,
          file.path(here::here(),
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_1x1_count_hooks.csv")),
          row.names = FALSE)

# Using just GFW reported effort (so can use all catch reported, regardless of effort)
combined_data <- short_data %>%
  group_by(rfmo, year, latitude, longitude, gear_group, species_group, spatial_notes, species_commonname, species_sciname, catch_units, species_resolution) %>%
  summarise(catch = sum(catch, na.rm = T)) %>%
  ungroup() %>%
  mutate(year = as.integer(year)) %>%
  filter(spatial_notes == "center of 1x1 cell" & catch_units == "count") %>%
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) %>% # lower limit for GFW data
  left_join(all_sdm_1x1 %>% distinct_all()) %>%
  left_join(gfw_data_1x1 %>% distinct_all()) %>%
  left_join(sst_data_1x1 %>% distinct_all()) %>%
  left_join(chla_data_1x1 %>% distinct_all()) %>%
  left_join(ssh_data_1x1 %>% distinct_all())

write.csv(combined_data,
          file.path(here::here(),
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_1x1_count_kwh.csv")),
          row.names = FALSE)

if(length(unique(short_data$catch_units)) > 1) {
# Update mt to count
# Run conversions
short_data_mt <- short_data %>%
  filter(catch_units == "metric tonnes") %>%
  left_join(mt_count, by = c("species_sciname" = "scientific_name")) %>%
  mutate(catch = catch/weight_mt,
         catch_units = "count") %>%
  select(colnames(short_data))

# Add back to dataset
short_data <- short_data %>%
  filter(catch_units == "count") %>%
  bind_rows(short_data_mt) %>%
  group_by_at(colnames(short_data)[!colnames(short_data) == "catch"]) %>%
  summarise(catch = sum(catch, na.rm = T)) %>%
  ungroup()

short_data <- short_data %>%
  distinct_all()

# Combine the data
# Using just the self-reported effort by hooks (so can only use catch by hooks)
combined_data <- short_data %>%
  filter(effort_units == "hooks") %>%
  mutate(year = as.integer(year)) %>%
  filter(spatial_notes == "center of 1x1 cell" & catch_units == "count") %>%
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) %>% # lower limit for GFW data
  left_join(all_sdm_1x1 %>% distinct_all()) %>%
  left_join(sst_data_1x1 %>% distinct_all()) %>%
  left_join(chla_data_1x1 %>% distinct_all()) %>%
  left_join(ssh_data_1x1 %>% distinct_all())

write.csv(combined_data,
          file.path(here::here(),
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_1x1_mt_to_count_hooks.csv")),
          row.names = FALSE)

# Using just GFW reported effort (so can use all catch reported, regardless of effort)
combined_data <- short_data %>%
  group_by(rfmo, year, latitude, longitude, gear_group, species_group, spatial_notes, species_commonname, species_sciname, catch_units, species_resolution) %>%
  summarise(catch = sum(catch, na.rm = T)) %>%
  ungroup() %>%
  mutate(year = as.integer(year)) %>%
  filter(spatial_notes == "center of 1x1 cell" & catch_units == "count") %>%
  filter(year >= min_yr & year <= max_yr$year[which(max_yr$rfmo == rfmos)]) %>% # lower limit for GFW data
  left_join(all_sdm_1x1 %>% distinct_all()) %>%
  left_join(gfw_data_1x1 %>% distinct_all()) %>%
  left_join(sst_data_1x1 %>% distinct_all()) %>%
  left_join(chla_data_1x1 %>% distinct_all()) %>%
  left_join(ssh_data_1x1 %>% distinct_all())

write.csv(combined_data,
          file.path(here::here(),
          paste0("data-updated/model-data/inputs/all-rfmo-models/", rfmos, "_ll_data_1x1_mt_to_count_kwh.csv")),
          row.names = FALSE)
}
} 
```
